# apps/backend/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "SaveMate API",
    "version": "0.1.0"
  },
  "servers": [
    {
      "url": "http://localhost:4000"
    }
  ],
...
  "paths": {
...
    "/admin/deals/pending": {
      "get": {
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1
            },
            "required": false,
            "name": "page",
            "in": "query"
          },
          {
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 50,
              "default": 20
            },
            "required": false,
            "name": "limit",
            "in": "query"
          },
          {
            "schema": {
              "type": "string",
              "enum": [
                "newest"
              ],
              "default": "newest"
            },
            "required": false,
            "name": "sort",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "items": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "string",
                            "minLength": 1
                          },
                          "title": {
                            "type": "string",
                            "minLength": 1
                          },
                          "price": {
                            "type": "number",
                            "minimum": 0
                          },
                          "originalPrice": {
                            "type": "number",
                            "minimum": 0
                          },
                          "discountPercent": {
                            "type": "integer",
                            "minimum": 0,
                            "maximum": 100
                          },
                          "status": {
                            "type": "string",
                            "enum": [
                              "DRAFT",
                              "PENDING",
                              "APPROVED",
                              "REJECTED",
                              "EXPIRED"
                            ]
                          },
                          "city": {
                            "type": "string",
                            "minLength": 1
                          },
                          "voivodeship": {
                            "type": "string",
                            "enum": [
                              "DOLNOSLASKIE",
                              "KUJAWSKO_POMORSKIE",
                              "LUBELSKIE",
                              "LUBUSKIE",
                              "LODZKIE",
                              "MALOPOLSKIE",
                              "MAZOWIECKIE",
                              "OPOLSKIE",
                              "PODKARPACKIE",
                              "PODLASKIE",
                              "POMORSKIE",
                              "SLASKIE",
                              "SWIETOKRZYSKIE",
                              "WARMINSKO_MAZURSKIE",
                              "WIELKOPOLSKIE",
                              "ZACHODNIOPOMORSKIE"
                            ]
                          },
                          "categoryId": {
                            "type": "string",
                            "minLength": 1
                          },
                          "tags": {
                            "type": "array",
                            "items": {
                              "type": "string"
                            }
                          },
                          "validFrom": {
                            "type": "string",
                            "format": "date-time"
                          },
                          "validTo": {
                            "type": "string",
                            "format": "date-time"
                          },
                          "createdAt": {
                            "type": "string",
                            "format": "date-time"
                          },
                          "businessId": {
                            "type": "string",
                            "minLength": 1
                          }
                        },
                        "required": [
                          "id",
                          "title",
                          "price",
                          "originalPrice",
                          "discountPercent",
                          "status",
                          "city",
                          "voivodeship",
                          "categoryId",
                          "tags",
                          "validFrom",
                          "validTo",
                          "createdAt",
                          "businessId"
                        ]
                      }
                    },
                    "page": {
                      "type": "object",
                      "properties": {
                        "page": {
                          "type": "integer",
                          "minimum": 1
                        },
                        "limit": {
                          "type": "integer",
                          "minimum": 1
                        },
                        "total": {
                          "type": "integer",
                          "minimum": 0
                        },
                        "totalPages": {
                          "type": "integer",
                          "minimum": 0
                        }
                      },
                      "required": [
                        "page",
                        "limit",
                        "total",
                        "totalPages"
                      ]
                    }
                  },
                  "required": [
                    "items",
                    "page"
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "object",
                      "properties": {
                        "code": {
                          "type": "string",
                          "enum": [
                            "VALIDATION_ERROR",
                            "UNAUTHORIZED",
                            "FORBIDDEN",
                            "NOT_FOUND",
                            "CONFLICT",
                            "RATE_LIMITED",
                            "INTERNAL"
                          ]
                        },
                        "message": {
                          "type": "string"
                        },
                        "details": {
                          "nullable": true
                        },
                        "requestId": {
                          "type": "string"
                        }
                      },
                      "required": [
                        "code",
                        "message",
                        "requestId"
                      ]
                    }
                  },
                  "required": [
                    "error"
                  ]
                }
              }
            }
          },
          "401": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "object",
                      "properties": {
                        "code": {
                          "type": "string",
                          "enum": [
                            "VALIDATION_ERROR",
                            "UNAUTHORIZED",
                            "FORBIDDEN",
                            "NOT_FOUND",
                            "CONFLICT",
                            "RATE_LIMITED",
                            "INTERNAL"
                          ]
                        },
                        "message": {
                          "type": "string"
                        },
                        "details": {
                          "nullable": true
                        },
                        "requestId": {
                          "type": "string"
                        }
                      },
                      "required": [
                        "code",
                        "message",
                        "requestId"
                      ]
                    }
                  },
                  "required": [
                    "error"
                  ]
                }
              }
            }
          },
          "403": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "object",
                      "properties": {
                        "code": {
                          "type": "string",
                          "enum": [
                            "VALIDATION_ERROR",
                            "UNAUTHORIZED",
                            "FORBIDDEN",
                            "NOT_FOUND",
                            "CONFLICT",
                            "RATE_LIMITED",
                            "INTERNAL"
                          ]
                        },
                        "message": {
                          "type": "string"
                        },
                        "details": {
                          "nullable": true
                        },
                        "requestId": {
                          "type": "string"
                        }
                      },
                      "required": [
                        "code",
                        "message",
                        "requestId"
                      ]
                    }
                  },
                  "required": [
                    "error"
                  ]
                }
              }
            }
          },
          "500": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "object",
                      "properties": {
                        "code": {
                          "type": "string",
                          "enum": [
                            "VALIDATION_ERROR",
                            "UNAUTHORIZED",
                            "FORBIDDEN",
                            "NOT_FOUND",
                            "CONFLICT",
                            "RATE_LIMITED",
                            "INTERNAL"
                          ]
                        },
                        "message": {
                          "type": "string"
                        },
                        "details": {
                          "nullable": true
                        },
                        "requestId": {
                          "type": "string"
                        }
                      },
                      "required": [
                        "code",
                        "message",
                        "requestId"
                      ]
                    }
                  },
                  "required": [
                    "error"
                  ]
                }
              }
            }
          }
        }
      }
    },
    "/admin/deals/{id}/approve": {
      "post": {
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "minLength": 1
            },
            "required": true,
            "name": "id",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string",
                      "minLength": 1
                    },
                    "status": {
                      "type": "string",
                      "enum": [
                        "PENDING",
                        "DRAFT",
                        "APPROVED",
                        "REJECTED",
                        "EXPIRED"
                      ]
                    }
                  },
                  "required": [
                    "id",
                    "status"
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "object",
                      "properties": {
                        "code": {
                          "type": "string",
                          "enum": [
                            "VALIDATION_ERROR",
                            "UNAUTHORIZED",
                            "FORBIDDEN",
                            "NOT_FOUND",
                            "CONFLICT",
                            "RATE_LIMITED",
                            "INTERNAL"
                          ]
                        },
                        "message": {
                          "type": "string"
                        },
                        "details": {
                          "nullable": true
                        },
                        "requestId": {
                          "type": "string"
                        }
                      },
                      "required": [
                        "code",
                        "message",
                        "requestId"
                      ]
                    }
                  },
                  "required": [
                    "error"
                  ]
                }
              }
            }
          },
          "401": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "object",
                      "properties": {
                        "code": {
                          "type": "string",
                          "enum": [
                            "VALIDATION_ERROR",
                            "UNAUTHORIZED",
                            "FORBIDDEN",
                            "NOT_FOUND",
                            "CONFLICT",
                            "RATE_LIMITED",
                            "INTERNAL"
                          ]
                        },
                        "message": {
                          "type": "string"
                        },
                        "details": {
                          "nullable": true
                        },
                        "requestId": {
                          "type": "string"
                        }
                      },
                      "required": [
                        "code",
                        "message",
                        "requestId"
                      ]
                    }
                  },
                  "required": [
                    "error"
                  ]
                }
              }
            }
          },
          "403": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "object",
                      "properties": {
                        "code": {
                          "type": "string",
                          "enum": [
                            "VALIDATION_ERROR",
                            "UNAUTHORIZED",
                            "FORBIDDEN",
                            "NOT_FOUND",
                            "CONFLICT",
                            "RATE_LIMITED",
                            "INTERNAL"
                          ]
                        },
                        "message": {
                          "type": "string"
                        },
                        "details": {
                          "nullable": true
                        },
                        "requestId": {
                          "type": "string"
                        }
                      },
                      "required": [
                        "code",
                        "message",
                        "requestId"
                      ]
                    }
                  },
                  "required": [
                    "error"
                  ]
                }
              }
            }
          },
          "404": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "object",
                      "properties": {
                        "code": {
                          "type": "string",
                          "enum": [
                            "VALIDATION_ERROR",
                            "UNAUTHORIZED",
                            "FORBIDDEN",
                            "NOT_FOUND",
                            "CONFLICT",
                            "RATE_LIMITED",
                            "INTERNAL"
                          ]
                        },
                        "message": {
                          "type": "string"
                        },
                        "details": {
                          "nullable": true
                        },
                        "requestId": {
                          "type": "string"
                        }
                      },
                      "required": [
                        "code",
                        "message",
                        "requestId"
                      ]
                    }
                  },
                  "required": [
                    "error"
                  ]
                }
              }
            }
          },
          "409": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "object",
                      "properties": {
                        "code": {
                          "type": "string",
                          "enum": [
                            "VALIDATION_ERROR",
                            "UNAUTHORIZED",
                            "FORBIDDEN",
                            "NOT_FOUND",
                            "CONFLICT",
                            "RATE_LIMITED",
                            "INTERNAL"
                          ]
                        },
                        "message": {
                          "type": "string"
                        },
                        "details": {
                          "nullable": true
                        },
                        "requestId": {
                          "type": "string"
                        }
                      },
                      "required": [
                        "code",
                        "message",
                        "requestId"
                      ]
                    }
                  },
                  "required": [
                    "error"
                  ]
                }
              }
            }
          },
          "500": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "object",
                      "properties": {
                        "code": {
                          "type": "string",
                          "enum": [
                            "VALIDATION_ERROR",
                            "UNAUTHORIZED",
                            "FORBIDDEN",
                            "NOT_FOUND",
                            "CONFLICT",
                            "RATE_LIMITED",
                            "INTERNAL"
                          ]
                        },
                        "message": {
                          "type": "string"
                        },
                        "details": {
                          "nullable": true
                        },
                        "requestId": {
                          "type": "string"
                        }
                      },
                      "required": [
                        "code",
                        "message",
                        "requestId"
                      ]
                    }
                  },
                  "required": [
                    "error"
                  ]
                }
              }
            }
          }
        }
      }
    },
...
  }
}